requirements.txt
aiofiles==0.7.0
aiosqlite==0.16.1
asyncpg==0.23.0
click==7.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
iso8601==0.1.14
multipart==0.2.4
pydantic==1.8.2
pypika-tortoise==0.1.1
python-dotenv==0.17.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
starlette==0.14.2
tortoise-orm==0.17.3
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
